pyproject.toml
[project]
name = "open-social-poster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "atproto>=0.0.59",
    "mastodon-py>=2.0.0",
    "monsterui>=1.0.10",
    "pillow>=11.1.0",
    "pyinstaller>=6.12.0",
    "pystray>=0.19.5",
    "python-fasthtml>=0.12.1",
    "selenium>=4.29.0",
]
